Data on the Cultural Dimension of Urban Community Food Consumption in Indonesia
This dataset shows survey data on the characteristics of the cultural dimension of food consumption patterns in urban communities in Indonesia. The cultural dimension discussed in this article includes the cultural value and cultural practice sub-dimensions. Data is enriched by illustrating the cultural characteristics of urban people’s food consumption patterns based on social class categories. Data collection was conducted in five major cities of Indonesia with face-to-face interviews with 710 respondents obtained using a stratified random sampling technique.